Move InMemoryBackendStorage to separate module
