+++ backend/jmap/src/test/java/com/fsck/k9/backend/jmap/CommandRefreshFolderListTest.kt
@@ -1,5 +1,6 @@
 package com.fsck.k9.backend.jmap
 
+import app.k9mail.backend.testing.InMemoryBackendStorage
 import com.fsck.k9.backend.api.BackendFolderUpdater
 import com.fsck.k9.backend.api.FolderInfo
 import com.fsck.k9.backend.api.updateFolders
