+++ settings.gradle
@@ -18,6 +18,7 @@ include ':mail:protocols:pop3'
 include ':mail:protocols:webdav'
 include ':mail:protocols:smtp'
 include ':backend:api'
+include ':backend:testing'
 include ':backend:imap'
 include ':backend:pop3'
 include ':backend:webdav'
